+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 15-MAR-2026 CEST
+Last change: 16-MAR-2026 CEST
 
 0.10.0
 ------
@@ -21,7 +21,8 @@ Bug fixes:
 Deprecation warnings:
 +++++++++++++++++++++
 - ``picasso.lib.unpack_calibration`` and the ``spot_size``, ``z_range`` parameters in the G5M functions. ``picasso.g5m.g5m`` now uses calibration coefficients only for setting sigma constraints in 3D for more accurate results.
-- ``picasso.clusterer.cluster_center`` (will be renamed to ``_cluster_center`` and become a private function in the future release)
+- ``picasso.clusterer.cluster_center`` (will be renamed to ``_cluster_center`` and become a private function in v0.11.0)
+- ``picasso.aim``: ``intersect1d``, ``count_intersections``, ``run_intersections``, ``run_intersections_multithread``, ``get_fft_peak``, ``get_fft_peak_z``, ``point_intersect_2d`` and ``point_intersect_3d`` (will become private functions in v0.11.0)
 
 0.9.8-9
 -------